% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch Application}
\usage{
launch_app(app = "explorer", use_browser = TRUE)
}
\arguments{
\item{app}{Name of the application.}

\item{use_browser}{Logical.}
}
\description{
Launch Application
}
\examples{
\dontrun{launch_app(app = "explorer", user_browser = TRUE)}

}
